+++ src/adk_interviewer/tools/__init__.py
@@ -3,10 +3,14 @@ from .question_generator import generate_question
 from .answer_evaluator import evaluate_answer
 from .resume_parser import parse_resume
 from .jd_analyzer import analyze_job_description
+from .concept_explainer import explain_concept
+from .hint_provider import provide_hints
 
 __all__ = [
     "generate_question",
     "evaluate_answer",
     "parse_resume",
-    "analyze_job_description"
+    "analyze_job_description",
+    "explain_concept",
+    "provide_hints"
 ]